requirements.txt
Flask==3.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-migrate==4.0.5
flask-limiter==3.5.0
anthropic==0.26.1
Pillow==10.2.0
python-dotenv==1.0.0
requests==2.31.0
twilio==8.13.0
reportlab==4.0.7
APScheduler==3.10.4
gunicorn==21.2.0
psycopg2-binary==2.9.9
redis==5.0.1
werkzeug==3.0.1